Imtihon/requirements.txt
Django==5.1.6
django-restframework==0.0.1
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-yasg==1.21.8
psycopg2==2.9.10
pillow==11.0.0